Add changes to update-version.sh

+++ update-version.sh
@@ -1,10 +1,19 @@
 #!/usr/bin/env bash
 
+if [ "$1" == "push" ]; then
+	cd ~/WordPress/wplib-wordpress/
+	echo "Push new version...";
+	git push
+	git push --tags
+	exit
+fi;
+
 if [ "$1" != "" ]; then
 	cd ~/WordPress/wplib-wordpress/
 	if [ ! -d "../wordpress-$1" ]; then
 		echo "There is no directory ${PWD%/wplib-wordpress}/wordpress-${1}.";
 	else
+		echo "Deleting prior version files...";
 		rm *.php
 		rm *.html
 		rm *.txt
@@ -12,6 +21,7 @@ if [ "$1" != "" ]; then
 		rm -Rf wp-*
 		mkdir log
 		git checkout master
+		git reset --hard
 		echo "Copying wordpress-${1}...";
 		cp -r ../wordpress-$1/.  .  > log/copy.log
 		git pull > log/pull.log
@@ -22,6 +32,3 @@ if [ "$1" != "" ]; then
 
 fi;
 
-if [ "$1" != "push" ]; then
-#		git push
-#		git push --tags
